requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
nltk==3.9.1
openai==1.65.3
pydantic==2.10.6
pydantic_core==2.27.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
